api/requirements.txt
asyncpg
logfire
pydantic
pydantic-ai
devtools
matplotlib
pandas
seaborn
numpy
fastapi
uvicorn[standard]
websockets
python-multipart
python-dotenv
pydantic_settings

